requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiohttp-proxy==0.1.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.6.2.post1
APScheduler==3.6.3
astunparse==1.6.3
asyncio-dgram==2.2.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
browser-cookie3==0.20.1
cachetools==4.2.2
certifi==2024.8.30
cfscrape==2.1.1
chardet==5.2.0
charset-normalizer==3.4.0
class-registry==2.1.2
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
DateTime==5.5
dnspython==2.7.0
dnspython3==1.15.0
EditorConfig==0.12.4
environs==14.1.1
escodegen==1.0.11
esprima==4.0.1
esutils==1.0.1
fake-useragent==2.0.3
ffmpeg==1.4
ffmpeg-python==0.2.0
filters==1.3.2
Flask==3.0.3
frozenlist==1.5.0
future==1.0.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
googlesearch-python==1.2.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jeepney==0.9.0
Jinja2==3.1.4
jsbeautifier==1.15.1
lz4==4.4.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcstatus==5.0.0
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.1.0
mutagen==1.47.0
nest-asyncio==1.6.0
outcome==1.3.0.post0
packaging==24.2
pexpect==4.9.0
pillow==11.0.0
propcache==0.2.0
psutil==7.0.0
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycryptodome==3.21.0
pycryptodomex==3.22.0
pyfiglet==1.0.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.0
pymongo==4.11.3
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
Pyrogram==2.0.106
PyScreeze==1.0.1
PySocks==1.7.1
pyTelegramBotAPI==4.18.1
pytgcalls==2.1.0
python-apt @ file:///home/builder/.termux-build/python-apt/src
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-dotenv==1.0.1
python-telegram-bot==22.0
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
pywhatkit==5.4
regex==2024.9.11
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.3
rsa==4.9
selenium==4.27.1
setuptools==78.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
socks==0
sortedcontainers==2.4.0
soupsieve==2.6
speedtest-cli==2.1.3
telebot==0.0.5
telegram==0.0.1
Telethon==1.39.0
termcolor==2.5.0
TgCrypto==1.2.5
tomli==2.0.2
tornado==6.4.1
trio==0.28.0
trio-websocket==0.11.1
twilio==9.3.7
typer==0.12.5
typing_extensions==4.12.2
tzlocal==5.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
validation==0.8.3
waitress==3.0.2
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.0
wheel==0.45.1
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.17.2
zope.interface==7.1.1
